requirements.txt
requests
python-dotenv
pypdf
openai
openai-agents